livekit-agent/requirements.txt
livekit>=0.11.0
livekit-agents>=0.8.0
livekit-plugins-openai>=0.6.0
python-dotenv>=1.0.0
